requirements.txt
annotated-types==0.7.0
appengine-python-standard==1.1.10
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
flask==3.1.2
frozendict==2.4.6
google-auth==2.43.0
icalendar==6.3.2
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
legacy-cgi==2.6.4
markupsafe==3.0.3
mock==5.2.0
pillow==12.0.0
protobuf==6.33.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.4
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.14
six==1.17.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==1.26.20
werkzeug==3.1.3
